# spec/openapi-v1.5.yaml
openapi: 3.1.0
info:
  title: Requirements Version Management System API
  version: 1.5.0
  description: 独立部署版 V1.5 完整开发基线接口契约。所有关键写操作通过 revision 乐观锁避免多人编辑静默覆盖。
servers:
- url: /api/v1
...
paths:
...
  /versions:
    get:
      tags:
      - Version
      summary: 版本列表
      security: *id001
      parameters: *id002
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
    post:
      tags:
      - Version
      summary: 创建版本
      security: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCreate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
  /versions/{version_id}:
    get:
      tags:
      - Version
      summary: 版本详情
      security: *id001
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
    patch:
      tags:
      - Version
      summary: 编辑版本
      security: *id001
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionUpdate'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
        '409':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
  /versions/{version_id}/status:
    patch:
      tags:
      - Version
      summary: 变更版本状态
      security: *id001
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionStatusChange'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Version'
  /versions/{version_id}/publish:
    post:
      tags:
      - Version
      summary: 发布版本
      description: 只允许 READY。Release + Version + DONE Requirement -> ONLINE + Feedback -> ONLINE 同一业务事务。
      security: *id001
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishVersionRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
        '409':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VersionCreate:
      type: object
      required:
      - version_no
      - name
      properties:
        version_no:
          type: string
        name:
          type: string
        owner_id:
          type:
          - integer
          - 'null'
        planned_release_date:
          type:
          - string
          - 'null'
          format: date
        description:
          type:
          - string
          - 'null'
    VersionUpdate:
      type: object
      required:
      - revision
      properties:
        name:
          type: string
        owner_id:
          type:
          - integer
          - 'null'
        planned_release_date:
          type:
          - string
          - 'null'
          format: date
        description:
          type:
          - string
          - 'null'
        revision:
          type: integer
          minimum: 1
    VersionStatusChange:
      type: object
      required:
      - status
      - revision
      properties:
        status:
          type: string
          enum:
          - PLANNING
          - DEVELOPING
          - TESTING
          - READY
          - RELEASED
          - CANCELED
        revision:
          type: integer
    PublishVersionRequest:
      type: object
      required:
      - released_at
      - result
      - release_notes
      - revision
      properties:
        released_at:
          type: string
          format: date-time
        result:
          type: string
          enum:
          - SUCCESS
          - PARTIAL
          - ROLLED_BACK
        release_notes:
          type: string
        revision:
          type: integer
          minimum: 1
    Version:
      type: object
      properties:
        id:
          type: integer
        version_no:
          type: string
        name:
          type: string
        status:
          type: string
        owner_id:
          type:
          - integer
          - 'null'
        planned_release_date:
          type:
          - string
          - 'null'
          format: date
        released_at:
          type:
          - string
          - 'null'
          format: date-time
        description:
          type:
          - string
          - 'null'
        revision:
          type: integer
...
    Conflict:
      type: object
      properties:
        code:
          type: integer
          example: 40910
        message:
          type: string
        data:
          type: object
          properties:
            current_revision:
              type: integer
            current_updated_at:
              type: string
            current_updated_by:
              type: integer
        request_id:
          type: string